# perf(tea): inlining encode