```css
name         old time/op    new time/op    delta
TEAen/16-8      253ns ± 1%     247ns ± 1%  -2.18%  (p=0.008 n=5+5)
TEAen/256-8    1.77µs ± 1%    1.72µs ± 0%  -3.31%  (p=0.008 n=5+5)
TEAen/4K-8     26.0µs ± 1%    25.2µs ± 1%  -2.91%  (p=0.008 n=5+5)
TEAen/32K-8     210µs ± 0%     204µs ± 0%  -2.97%  (p=0.008 n=5+5)
TEAde/16-8      217ns ± 0%     211ns ± 0%  -2.51%  (p=0.008 n=5+5)
TEAde/256-8    1.70µs ± 0%    1.66µs ± 0%  -2.90%  (p=0.008 n=5+5)
TEAde/4K-8     25.5µs ± 1%    24.8µs ± 0%  -2.61%  (p=0.008 n=5+5)
TEAde/32K-8     206µs ± 0%     200µs ± 0%  -2.56%  (p=0.008 n=5+5)

name         old speed      new speed      delta
TEAen/16-8   63.3MB/s ± 1%  64.7MB/s ± 1%  +2.22%  (p=0.008 n=5+5)
TEAen/256-8   144MB/s ± 1%   149MB/s ± 0%  +3.43%  (p=0.008 n=5+5)
TEAen/4K-8    158MB/s ± 1%   162MB/s ± 1%  +2.99%  (p=0.008 n=5+5)
TEAen/32K-8   156MB/s ± 0%   161MB/s ± 0%  +3.06%  (p=0.008 n=5+5)
TEAde/16-8    148MB/s ± 0%   151MB/s ± 0%  +2.58%  (p=0.008 n=5+5)
TEAde/256-8   160MB/s ± 0%   164MB/s ± 0%  +2.99%  (p=0.008 n=5+5)
TEAde/4K-8    161MB/s ± 1%   166MB/s ± 0%  +2.68%  (p=0.008 n=5+5)
TEAde/32K-8   159MB/s ± 0%   164MB/s ± 0%  +2.63%  (p=0.008 n=5+5)
```